gall + agents: add breached flag to +on-load
Changes the type of $agent:gall to include a breached flag in +on-load.
Updates agents to conform to this type.

If the breached flag is %.y, then all vane state has been wiped and must
be recreated. Most agents can simply ignore this flag. See
link-listen-hook for an example of recreating the vane state.